+++ kotlin/gltfio/src/androidMain/kotlin/io/github/erkko68/filament/gltfio/MaterialProvider.android.kt
@@ -4,12 +4,13 @@ import io.github.erkko68.filament.Engine
 import io.github.erkko68.filament.FilamentPlatform
 import io.github.erkko68.filament.PlatformGap
 import io.github.erkko68.filament.nativeObject
+import io.github.erkko68.filament.VertexBuffer
 
 actual interface MaterialProvider : AutoCloseable {
     actual fun createMaterialInstance(config: MaterialKey, uvmap: IntArray, label: String?, extras: String?): io.github.erkko68.filament.MaterialInstance?
     actual fun getMaterial(config: MaterialKey, uvmap: IntArray, label: String?): io.github.erkko68.filament.Material?
     actual val materials: List<io.github.erkko68.filament.Material>
-    actual fun needsDummyData(attrib: Int): Boolean
+    actual fun needsDummyData(attrib: VertexBuffer.VertexAttribute): Boolean
     actual fun destroyMaterials()
     actual fun destroy()
     
@@ -21,12 +22,12 @@ actual class UbershaderProvider actual constructor(engine: Engine) : MaterialPro
     private val nativeObject = com.google.android.filament.gltfio.UbershaderProvider(engine.nativeObject)
 
     actual override fun createMaterialInstance(config: MaterialKey, uvmap: IntArray, label: String?, extras: String?): io.github.erkko68.filament.MaterialInstance? {
-        val nativeInstance = nativeObject.createMaterialInstance(config.nativeObject, uvmap, label, extras) ?: return null
+        val nativeInstance = nativeObject.createMaterialInstance(config.toAndroid(), uvmap, label, extras) ?: return null
         return io.github.erkko68.filament.MaterialInstance(nativeInstance)
     }
 
     actual override fun getMaterial(config: MaterialKey, uvmap: IntArray, label: String?): io.github.erkko68.filament.Material? {
-        val nativeMaterial = nativeObject.getMaterial(config.nativeObject, uvmap, label) ?: return null
+        val nativeMaterial = nativeObject.getMaterial(config.toAndroid(), uvmap, label) ?: return null
         return io.github.erkko68.filament.Material(nativeMaterial)
     }
 
@@ -35,7 +36,7 @@ actual class UbershaderProvider actual constructor(engine: Engine) : MaterialPro
         return List(natives.size) { i -> io.github.erkko68.filament.Material(natives[i]) }
     }
 
-    actual override fun needsDummyData(attrib: Int): Boolean = nativeObject.needsDummyData(attrib)
+    actual override fun needsDummyData(attrib: VertexBuffer.VertexAttribute): Boolean = nativeObject.needsDummyData(attrib.ordinal)
 
     actual override fun destroyMaterials() {
         nativeObject.destroyMaterials()