+++ kotlin/gltfio/src/jvmMain/kotlin/io/github/erkko68/filament/gltfio/MaterialProvider.jvm.kt
@@ -15,12 +15,13 @@ import io.github.erkko68.filament.FilamentPlatform
 import io.github.erkko68.filament.PlatformGap
 import io.github.erkko68.filament.InternalFilamentApi
 import io.github.erkko68.filament.nativeObject
+import io.github.erkko68.filament.VertexBuffer
 
 actual interface MaterialProvider : AutoCloseable {
     actual fun createMaterialInstance(config: MaterialKey, uvmap: IntArray, label: String?, extras: String?): MaterialInstance?
     actual fun getMaterial(config: MaterialKey, uvmap: IntArray, label: String?): Material?
     actual val materials: List<Material>
-    actual fun needsDummyData(attrib: Int): Boolean
+    actual fun needsDummyData(attrib: VertexBuffer.VertexAttribute): Boolean
     actual fun destroyMaterials()
     actual fun destroy()
 
@@ -66,7 +67,7 @@ actual class UbershaderProvider actual constructor(engine: Engine) : MaterialPro
         }
     }
 
-    actual override fun needsDummyData(attrib: Int): Boolean = FilamentC.FilaMaterialProvider_needsDummyData(nativeHandle, attrib)
+    actual override fun needsDummyData(attrib: VertexBuffer.VertexAttribute): Boolean = FilamentC.FilaMaterialProvider_needsDummyData(nativeHandle, attrib.ordinal)
     actual override fun destroyMaterials() = FilamentC.FilaMaterialProvider_destroyMaterials(nativeHandle)
 
     actual override fun close() = destroy()